streamlit/requirements.txt
streamlit
streamlit-extras
boto3
requests
python-dotenv
openai==0.28